SectionPage: scroll to top when navigating to new page
