+++ client/pages/Assessment/Assessment.tsx
@@ -1,18 +1,12 @@
 import './Assessment.scss'
-import React, { useEffect } from 'react'
-import { useParams } from 'react-router'
+import React from 'react'
 import { Route, Switch } from 'react-router-dom'
 
 import { Areas } from '@meta/area'
-import { AssessmentName } from '@meta/assessment'
 
-import { useAppDispatch } from '@client/store'
 import { useAssessment } from '@client/store/assessment'
-import { AssessmentSectionActions } from '@client/store/pages/assessmentSection'
 import { useNavigationVisible } from '@client/store/ui/navigation'
-import { ReviewActions } from '@client/store/ui/review'
-import { useUser } from '@client/store/user'
-import { useCountryIso, useOnUpdate } from '@client/hooks'
+import { useCountryIso } from '@client/hooks'
 import { BasePaths } from '@client/basePaths'
 import MessageCenter from '@client/components/MessageCenter'
 import Navigation from '@client/components/Navigation'
@@ -20,54 +14,7 @@ import AssessmentSection from '@client/pages/AssessmentSection'
 import DataExport from '@client/pages/DataExport'
 import OriginalDataPoint from '@client/pages/OriginalDataPoint'
 
-const SectionWrapper: React.FC = (props) => {
-  const { children } = props
-
-  const dispatch = useAppDispatch()
-  const countryIso = useCountryIso()
-  const user = useUser()
-  const { assessmentName, cycleName, section } = useParams<{
-    assessmentName: AssessmentName
-    cycleName: string
-    section: string
-  }>()
-
-  // fetch table sections metadata
-  useEffect(() => {
-    dispatch(
-      AssessmentSectionActions.getTableSections({
-        assessmentName,
-        cycleName,
-        section,
-        countryIso,
-      })
-    )
-  }, [countryIso, assessmentName, cycleName, section])
-
-  // fetch section review status
-  useEffect(() => {
-    if (user) {
-      dispatch(ReviewActions.getReviewStatus({ countryIso, assessmentName, cycleName, section }))
-    }
-  }, [countryIso, assessmentName, cycleName, section, user])
-
-  // fetch section summary
-  useEffect(() => {
-    if (user) {
-      dispatch(ReviewActions.getReviewSummary({ countryIso, assessmentName, cycleName }))
-    }
-  }, [countryIso, assessmentName, cycleName, user])
-
-  // reset store
-  useOnUpdate(() => {
-    return () => {
-      dispatch(AssessmentSectionActions.reset())
-      dispatch(ReviewActions.reset())
-    }
-  }, [countryIso, assessmentName, cycleName])
-
-  return <>{React.Children.toArray(children)}</>
-}
+import SectionWrapper from './SectionWrapper'
 
 const Assessment: React.FC = () => {
   const navigationVisible = useNavigationVisible()
