+++ client/pages/Assessment/SectionWrapper/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SectionWrapper'
